svg refactoring for v1_1, still not compiling
