+++ src/gui_interactive_controller.h
@@ -6,7 +6,7 @@ class GameController; // to be removed! ps_map is private, this is needed for fr
 
 #include "world_generator.h"
 #include "color_generator.h"
-#include "tobor_svg.h"
+#include "svg_1_1.h"
 
 // #include "tobor_svg.h" produces error
